- refactor and move itemsdisplay
