+++ OnlineMediaShelf.Web/clientapp/src/components/ItemsDisplay/ItemCard.tsx
